Cleanup and reorganize flags
