+++ beacon-chain/flags/BUILD.bazel
@@ -5,12 +5,13 @@ go_library(
     srcs = [
         "base.go",
         "config.go",
+        "grpc.go",
         "interop.go",
+        "p2p.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/flags",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
-        "//shared/cmd:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli_v2//:go_default_library",
     ],
