+++ beacon-chain/flags/base.go
@@ -7,6 +7,8 @@ import (
 )
 
 var (
+	// Network-related flags.
+
 	// HTTPWeb3ProviderFlag provides an HTTP access endpoint to an ETH 1.0 RPC.
 	HTTPWeb3ProviderFlag = &cli.StringFlag{
 		Name:  "http-web3provider",
@@ -18,6 +20,12 @@ var (
 		Name:  "deposit-contract",
 		Usage: "Deposit contract address. Beacon chain node will listen logs coming from the deposit contract to determine when validator is eligible to participate.",
 		Value: "0x07b39F4fDE4A38bACe212b546dAc87C58DfE3fDC", // Medalla deposit contract address.
+	}
+	// MonitoringPortFlag defines the http port used to serve prometheus metrics.
+	MonitoringPortFlag = &cli.IntFlag{
+		Name:  "monitoring-port",
+		Usage: "Port used to listening and respond metrics for prometheus.",
+		Value: 8080,
 	}
 	// RPCHost defines the host on which the RPC server should listen.
 	RPCHost = &cli.StringFlag{
@@ -31,59 +39,15 @@ var (
 		Usage: "RPC port exposed by a beacon node",
 		Value: 4000,
 	}
-	// MonitoringPortFlag defines the http port used to serve prometheus metrics.
-	MonitoringPortFlag = &cli.IntFlag{
-		Name:  "monitoring-port",
-		Usage: "Port used to listening and respond metrics for prometheus.",
-		Value: 8080,
-	}
-	// CertFlag defines a flag for the node's TLS certificate.
-	CertFlag = &cli.StringFlag{
-		Name:  "tls-cert",
-		Usage: "Certificate for secure gRPC. Pass this and the tls-key flag in order to use gRPC securely.",
-	}
-	// KeyFlag defines a flag for the node's TLS key.
-	KeyFlag = &cli.StringFlag{
-		Name:  "tls-key",
-		Usage: "Key for secure gRPC. Pass this and the tls-cert flag in order to use gRPC securely.",
-	}
-	// DisableGRPCGateway for JSON-HTTP requests to the beacon node.
-	DisableGRPCGateway = &cli.BoolFlag{
-		Name:  "disable-grpc-gateway",
-		Usage: "Disable the gRPC gateway for JSON-HTTP requests",
-	}
-	// GRPCGatewayHost specifies a gRPC gateway host for Prysm.
-	GRPCGatewayHost = &cli.StringFlag{
-		Name:  "grpc-gateway-host",
-		Usage: "The host on which the gateway server runs on",
-		Value: "127.0.0.1",
-	}
-	// GRPCGatewayPort enables a gRPC gateway to be exposed for Prysm.
-	GRPCGatewayPort = &cli.IntFlag{
-		Name:  "grpc-gateway-port",
-		Usage: "Enable gRPC gateway for JSON requests",
-		Value: 3500,
-	}
-	// GPRCGatewayCorsDomain serves preflight requests when serving gRPC JSON gateway.
-	GPRCGatewayCorsDomain = &cli.StringFlag{
-		Name: "grpc-gateway-corsdomain",
-		Usage: "Comma separated list of domains from which to accept cross origin requests " +
-			"(browser enforced). This flag has no effect if not used with --grpc-gateway-port.",
-		Value: "http://localhost:4242,http://127.0.0.1:4242,http://localhost:4200,http://0.0.0.0:4242,http://0.0.0.0:4200",
-	}
-	// MinSyncPeers specifies the required number of successful peer handshakes in order
-	// to start syncing with external peers.
-	MinSyncPeers = &cli.IntFlag{
-		Name:  "min-sync-peers",
-		Usage: "The required number of valid peers to connect with before syncing.",
-		Value: 3,
-	}
 	// ContractDeploymentBlock is the block in which the eth1 deposit contract was deployed.
 	ContractDeploymentBlock = &cli.IntFlag{
 		Name:  "contract-deployment-block",
 		Usage: "The eth1 block in which the deposit contract was deployed.",
 		Value: 11184524,
 	}
+
+	// Beacon node management flags.
+
 	// SetGCPercent is the percentage of current live allocations at which the garbage collector is to run.
 	SetGCPercent = &cli.IntFlag{
 		Name:  "gc-percent",
@@ -102,37 +66,18 @@ var (
 		Usage: "The slot durations of when an archived state gets saved in the DB.",
 		Value: 2048,
 	}
-	// DisableDiscv5 disables running discv5.
-	DisableDiscv5 = &cli.BoolFlag{
-		Name:  "disable-discv5",
-		Usage: "Does not run the discoveryV5 dht.",
-	}
-	// BlockBatchLimit specifies the requested block batch size.
-	BlockBatchLimit = &cli.IntFlag{
-		Name:  "block-batch-limit",
-		Usage: "The amount of blocks the local peer is bounded to request and respond to in a batch.",
-		Value: 64,
-	}
-	// BlockBatchLimitBurstFactor specifies the factor by which block batch size may increase.
-	BlockBatchLimitBurstFactor = &cli.IntFlag{
-		Name:  "block-batch-limit-burst-factor",
-		Usage: "The factor by which block batch limit may increase on burst.",
-		Value: 10,
+	// WeakSubjectivityCheckpt defines the weak subjectivity checkpoint the node must sync through to defend against long range attacks.
+	WeakSubjectivityCheckpt = &cli.StringFlag{
+		Name: "weak-subjectivity-checkpoint",
+		Usage: "Input in `block_root:epoch_number` format. This guarantee that syncing leads to the given Weak Subjectivity Checkpoint being in the canonical chain. " +
+			"If such a sync is not possible, the node will treat it a critical and irrecoverable failure",
+		Value: "",
 	}
 	// DisableSync disables a node from syncing at start-up. Instead the node enters regular sync
 	// immediately.
 	DisableSync = &cli.BoolFlag{
 		Name:  "disable-sync",
 		Usage: "Starts the beacon node without entering initial sync and instead exits to regular sync immediately.",
-	}
-	// EnableDebugRPCEndpoints as /v1/beacon/state.
-	EnableDebugRPCEndpoints = &cli.BoolFlag{
-		Name:  "enable-debug-rpc-endpoints",
-		Usage: "Enables the debug rpc service, containing utility endpoints such as /eth/v1alpha1/beacon/state.",
-	}
-	SubscribeToAllSubnets = &cli.BoolFlag{
-		Name:  "subscribe-all-subnets",
-		Usage: "Subscribe to all possible attestation subnets.",
 	}
 	// HistoricalSlasherNode is a set of beacon node flags required for performing historical detection with a slasher.
 	HistoricalSlasherNode = &cli.BoolFlag{
@@ -142,20 +87,16 @@ var (
 	// ChainID defines a flag to set the chain id. If none is set, it derives this value from NetworkConfig
 	ChainID = &cli.Uint64Flag{
 		Name:  "chain-id",
-		Usage: "Sets the chain id of the beacon chain",
+		Usage: "Sets the chain id of the beacon chain.",
 	}
 	// NetworkID defines a flag to set the network id. If none is set, it derives this value from NetworkConfig
 	NetworkID = &cli.Uint64Flag{
 		Name:  "network-id",
 		Usage: "Sets the network id of the beacon chain.",
 	}
-	// WeakSubjectivityCheckpt defines the weak subjectivity checkpoint the node must sync through to defend against long range attacks.
-	WeakSubjectivityCheckpt = &cli.StringFlag{
-		Name: "weak-subjectivity-checkpoint",
-		Usage: "Input in `block_root:epoch_number` format. This guarantee that syncing leads to the given Weak Subjectivity Checkpoint being in the canonical chain. " +
-			"If such a sync is not possible, the node will treat it a critical and irrecoverable failure",
-		Value: "",
-	}
+
+	// Backup flags.
+
 	// EnableBackupWebhookFlag for users to trigger db backups via an HTTP webhook.
 	EnableBackupWebhookFlag = &cli.BoolFlag{
 		Name:  "enable-db-backup-webhook",
