+++ beacon-chain/flags/p2p.go
@@ -1,0 +1,138 @@
+package flags
+
+import (
+	"github.com/urfave/cli/v2"
+)
+
+var (
+	// P2P management flags.
+
+	// MinSyncPeers specifies the required number of successful peer handshakes in order
+	// to start syncing with external peers.
+	MinSyncPeers = &cli.IntFlag{
+		Name:  "min-sync-peers",
+		Usage: "The required number of valid peers to connect with before syncing.",
+		Value: 3,
+	}
+	// DisableDiscv5 disables running discv5.
+	DisableDiscv5 = &cli.BoolFlag{
+		Name:  "disable-discv5",
+		Usage: "Does not run the discoveryV5 DHT.",
+	}
+	// SubscribeToAllSubnets subsribes the beacon node to every attestation subnet.
+	SubscribeToAllSubnets = &cli.BoolFlag{
+		Name:  "subscribe-all-subnets",
+		Usage: "Subscribe to all possible attestation subnets.",
+	}
+	// EnableUPnPFlag specifies if UPnP should be enabled or not. The default value is false.
+	EnableUPnPFlag = &cli.BoolFlag{
+		Name:  "enable-upnp",
+		Usage: "Enable the service (Beacon chain or Validator) to use UPnP when possible.",
+	}
+	// BlockBatchLimit specifies the requested block batch size.
+	BlockBatchLimit = &cli.IntFlag{
+		Name:  "block-batch-limit",
+		Usage: "The amount of blocks the local peer is bounded to request and respond to in a batch.",
+		Value: 64,
+	}
+	// BlockBatchLimitBurstFactor specifies the factor by which block batch size may increase.
+	BlockBatchLimitBurstFactor = &cli.IntFlag{
+		Name:  "block-batch-limit-burst-factor",
+		Usage: "The factor by which block batch limit may increase on burst.",
+		Value: 10,
+	}
+	// NoDiscovery specifies whether we are running a local network and have no need for connecting
+	// to the bootstrap nodes in the cloud
+	NoDiscovery = &cli.BoolFlag{
+		Name:  "no-discovery",
+		Usage: "Enable only local network p2p and do not connect to cloud bootstrap nodes.",
+	}
+	// StaticPeers specifies a set of peers to connect to explicitly.
+	StaticPeers = &cli.StringSliceFlag{
+		Name:  "peer",
+		Usage: "Connect with this peer. This flag may be used multiple times.",
+	}
+	// BootstrapNode tells the beacon node which bootstrap node to connect to
+	BootstrapNode = &cli.StringSliceFlag{
+		Name:  "bootstrap-node",
+		Usage: "The address of bootstrap node. Beacon node will connect for peer discovery via DHT.  Multiple nodes can be passed by using the flag multiple times but not comma-separated. You can also pass YAML files containing multiple nodes.",
+		Value: cli.NewStringSlice(
+			// Discv5
+			"enr:-Ku4QMKVC_MowDsmEa20d5uGjrChI0h8_KsKXDmgVQbIbngZV0idV6_RL7fEtZGo-kTNZ5o7_EJI_vCPJ6scrhwX0Z4Bh2F0dG5ldHOIAAAAAAAAAACEZXRoMpD1pf1CAAAAAP__________gmlkgnY0gmlwhBLf22SJc2VjcDI1NmsxoQJxCnE6v_x2ekgY_uoE1rtwzvGy40mq9eD66XfHPBWgIIN1ZHCCD6A",
+			// Discv5.1
+			"enr:-Ku4QOnVSyvzS3VbF87J8MubaRuTyfPi6B67XQg6-5eAV_uILAhn9geTTQmfqDIOcIeAxWHUUajQp6lYniAXPWncp6UBh2F0dG5ldHOIAAAAAAAAAACEZXRoMpAYrkzLAAAAAf__________gmlkgnY0gmlwhBLf22SJc2VjcDI1NmsxoQKekYKqUtwbaJKKCct_srE5-g7tBUm68mj_jpeSb7CCqYN1ZHCCC7g",
+		),
+	}
+	// RelayNode tells the beacon node which relay node to connect to.
+	RelayNode = &cli.StringFlag{
+		Name: "relay-node",
+		Usage: "The address of relay node. The beacon node will connect to the " +
+			"relay node and advertise their address via the relay node to other peers",
+		Value: "",
+	}
+
+	// P2P config flags.
+
+	// P2PUDPPort defines the port to be used by discv5.
+	P2PUDPPort = &cli.IntFlag{
+		Name:  "p2p-udp-port",
+		Usage: "The port used by discv5.",
+		Value: 12000,
+	}
+	// P2PTCPPort defines the port to be used by libp2p.
+	P2PTCPPort = &cli.IntFlag{
+		Name:  "p2p-tcp-port",
+		Usage: "The port used by libp2p.",
+		Value: 13000,
+	}
+	// P2PIP defines the local IP to be used by libp2p.
+	P2PIP = &cli.StringFlag{
+		Name:  "p2p-local-ip",
+		Usage: "The local ip address to listen for incoming data.",
+		Value: "",
+	}
+	// P2PHost defines the host IP to be used by libp2p.
+	P2PHost = &cli.StringFlag{
+		Name:  "p2p-host-ip",
+		Usage: "The IP address advertised by libp2p. This may be used to advertise an external IP.",
+		Value: "",
+	}
+	// P2PHostDNS defines the host DNS to be used by libp2p.
+	P2PHostDNS = &cli.StringFlag{
+		Name:  "p2p-host-dns",
+		Usage: "The DNS address advertised by libp2p. This may be used to advertise an external DNS.",
+		Value: "",
+	}
+	// P2PPrivKey defines a flag to specify the location of the private key file for libp2p.
+	P2PPrivKey = &cli.StringFlag{
+		Name:  "p2p-priv-key",
+		Usage: "The file containing the private key to use in communications with other peers.",
+		Value: "",
+	}
+	// P2PMetadata defines a flag to specify the location of the peer metadata file.
+	P2PMetadata = &cli.StringFlag{
+		Name:  "p2p-metadata",
+		Usage: "The file containing the metadata to communicate with other peers.",
+		Value: "",
+	}
+	// P2PMaxPeers defines a flag to specify the max number of peers in libp2p.
+	P2PMaxPeers = &cli.IntFlag{
+		Name:  "p2p-max-peers",
+		Usage: "The max number of p2p peers to maintain.",
+		Value: 30,
+	}
+	// P2PAllowList defines a CIDR subnet to exclusively allow connections.
+	P2PAllowList = &cli.StringFlag{
+		Name: "p2p-allowlist",
+		Usage: "The CIDR subnet for allowing only certain peer connections. Example: " +
+			"192.168.0.0/16 would permit connections to peers on your local network only. The " +
+			"default is to accept all connections.",
+	}
+	// P2PDenyList defines a list of CIDR subnets to disallow connections from them.
+	P2PDenyList = &cli.StringSliceFlag{
+		Name: "p2p-denylist",
+		Usage: "The CIDR subnets for denying certainy peer connections. Example: " +
+			"192.168.0.0/16 would deny connections from peers on your local network only. The " +
+			"default is to accept all connections.",
+	}
+)
