+++ beacon-chain/main.go
@@ -57,23 +57,23 @@ var appFlags = []cli.Flag{
 	flags.ChainID,
 	flags.NetworkID,
 	flags.WeakSubjectivityCheckpt,
+	flags.BootstrapNode,
+	flags.NoDiscovery,
+	flags.StaticPeers,
+	flags.RelayNode,
+	flags.P2PUDPPort,
+	flags.P2PTCPPort,
+	flags.P2PIP,
+	flags.P2PHost,
+	flags.P2PHostDNS,
+	flags.P2PMaxPeers,
+	flags.P2PPrivKey,
+	flags.P2PMetadata,
+	flags.P2PAllowList,
+	flags.P2PDenyList,
 	cmd.MinimalConfigFlag,
 	cmd.E2EConfigFlag,
 	cmd.RPCMaxPageSizeFlag,
-	cmd.BootstrapNode,
-	cmd.NoDiscovery,
-	cmd.StaticPeers,
-	cmd.RelayNode,
-	cmd.P2PUDPPort,
-	cmd.P2PTCPPort,
-	cmd.P2PIP,
-	cmd.P2PHost,
-	cmd.P2PHostDNS,
-	cmd.P2PMaxPeers,
-	cmd.P2PPrivKey,
-	cmd.P2PMetadata,
-	cmd.P2PAllowList,
-	cmd.P2PDenyList,
 	cmd.DataDirFlag,
 	cmd.VerbosityFlag,
 	cmd.EnableTracingFlag,
@@ -94,7 +94,7 @@ var appFlags = []cli.Flag{
 	debug.CPUProfileFlag,
 	debug.TraceFlag,
 	cmd.LogFileName,
-	cmd.EnableUPnPFlag,
+	flags.EnableUPnPFlag,
 	cmd.ConfigFileFlag,
 	cmd.ChainConfigFileFlag,
 	cmd.GrpcMaxCallRecvMsgSizeFlag,
