+++ beacon-chain/p2p/config.go
@@ -7,18 +7,18 @@ import (
 // Config for the p2p service. These parameters are set from application level flags
 // to initialize the p2p service.
 type Config struct {
+	DataDir             string
+	Discv5BootStrapAddr []string
 	NoDiscovery         bool
 	EnableUPnP          bool
 	DisableDiscv5       bool
 	StaticPeers         []string
 	BootstrapNodeAddr   []string
-	Discv5BootStrapAddr []string
 	RelayNodeAddr       string
 	LocalIP             string
 	HostAddress         string
 	HostDNS             string
 	PrivateKey          string
-	DataDir             string
 	MetaDataDir         string
 	TCPPort             uint
 	UDPPort             uint
