+++ beacon-chain/usage.go
@@ -47,11 +47,11 @@ var appHelpFlagGroups = []flagGroup{
 			cmd.MinimalConfigFlag,
 			cmd.E2EConfigFlag,
 			cmd.RPCMaxPageSizeFlag,
-			cmd.NoDiscovery,
-			cmd.BootstrapNode,
-			cmd.RelayNode,
-			cmd.P2PUDPPort,
-			cmd.P2PTCPPort,
+			flags.NoDiscovery,
+			flags.BootstrapNode,
+			flags.RelayNode,
+			flags.P2PUDPPort,
+			flags.P2PTCPPort,
 			cmd.DataDirFlag,
 			cmd.VerbosityFlag,
 			cmd.EnableTracingFlag,
@@ -117,16 +117,16 @@ var appHelpFlagGroups = []flagGroup{
 	{
 		Name: "p2p",
 		Flags: []cli.Flag{
-			cmd.P2PIP,
-			cmd.P2PHost,
-			cmd.P2PHostDNS,
-			cmd.P2PMaxPeers,
-			cmd.P2PPrivKey,
-			cmd.P2PMetadata,
-			cmd.P2PAllowList,
-			cmd.P2PDenyList,
-			cmd.StaticPeers,
-			cmd.EnableUPnPFlag,
+			flags.P2PIP,
+			flags.P2PHost,
+			flags.P2PHostDNS,
+			flags.P2PMaxPeers,
+			flags.P2PPrivKey,
+			flags.P2PMetadata,
+			flags.P2PAllowList,
+			flags.P2PDenyList,
+			flags.StaticPeers,
+			flags.EnableUPnPFlag,
 			flags.MinSyncPeers,
 		},
 	},
