+++ validator/flags/BUILD.bazel
@@ -3,13 +3,16 @@ load("@prysm//tools/go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
-        "flags.go",
+        "account.go",
+        "base.go",
+        "grpc.go",
         "interop.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/validator/flags",
     visibility = ["//validator:__subpackages__"],
     deps = [
         "//shared/fileutil:go_default_library",
+        "@com_github_urfave_cli//:go_default_library",
         "@com_github_urfave_cli_v2//:go_default_library",
     ],
 )
